+++ src/engraving/dom/stafftype.h
@@ -109,27 +109,8 @@ struct TablatureDurationFont {
     Spatium gridStemWidth  = GRID_STEM_DEF_WIDTH;       // the width of the 'grid'-style stem (in sp)
     // the note value with no beaming in 'grid'-style beaming
     DurationType zeroBeamLevel = DurationType::V_QUARTER;
-    Char displayDot = 'l';                            // the char to use to draw a dot
+    Char displayDot = u'l';                           // the char to use to draw a dot
     Char displayValue[int(TabVal::NUM_OF)];           // the char to use to draw a duration value
-};
-
-// ready-made staff types
-// keep in sync with the _presets initialization in StaffType::initStaffTypes() and _defaultPreset
-
-enum class StaffTypes : signed char {
-    STANDARD,
-    PERC_1LINE, PERC_2LINE, PERC_3LINE, PERC_5LINE,
-    TAB_6SIMPLE, TAB_6COMMON, TAB_6FULL,
-    TAB_4SIMPLE, TAB_4COMMON, TAB_4FULL,
-    TAB_5SIMPLE, TAB_5COMMON, TAB_5FULL,
-    TAB_UKULELE, TAB_BALALAJKA, TAB_DULCIMER,
-    TAB_ITALIAN, TAB_FRENCH,
-    TAB_7COMMON, TAB_8COMMON, TAB_9COMMON, TAB_10COMMON,
-    TAB_7SIMPLE, TAB_8SIMPLE, TAB_9SIMPLE, TAB_10SIMPLE,
-    STAFF_TYPES,
-    // some useful shorthands:
-    PERC_DEFAULT = StaffTypes::PERC_5LINE,
-    TAB_DEFAULT = StaffTypes::TAB_6COMMON,
 };
 
 enum class ShowTiedFret : unsigned char {
